# [opt](Nereids) restore the set_var hint after finish the execution